Support PHP8.1
This commit updates several key dependencies in the composer.json file, including the minimum PHP version, which is now 8.1.0, godruoyi/ocr library to version 3.0, and the illuminate/support package to version 10.0 for improved functionality and performance.

In the OCR.php file, the return type hints for the aliyun(), baidu(), and tencent() methods are modified for better readability, removing the full namespace path. This makes the code easier to understand without losing accuracy of information.